osd: avoid the config's get_val() overhead on the read path.
Profiling shows the overhead of the md_config_t::get_val<T>
can be significant. Unfortunately, it is being used in two
critical places across the read path.

The patch resorts to the legacy config infrastructure to
mitigate the performance penalty. It is expected it will
be superseded with a solution that allows to use the new,
intended routines without hurting performance.
`
Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>