include/any.h: Type Erasure for the uncopyable
std:any is nice and all, but they don't support uncopyable
objects. This is annoying to us, since we have a perfect use case for
std::any in the associated object map, but almost everything we want
to put into it is uncopyable due to having a mutex or something.

Enter ceph::immobile_any<>. A zero overhead type erasure that cannot
be moved or copied (you stick it in a container and it stays there) of
fixed size that does no allocation.

Also unique_any and shared_any, because occasionally move and copy
semantics are useful but you still want to be able to store an
arbitrary type.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>