osd: min_pg_log_entries == max == pg_log_dups_tracked
- avoid a memory utilization explosion when a pg is degraded
- avoid overhead of converting pg log entries to dup op entries.

The net of this is that during recovery we will keep fewer entries, which
means we will trigger backfill sooner rather than later.

In the future, it would still be nice to dynamically adjust this in such
a way as to avoid increasing the memory footprint (e.g., by borrowing
memory from BlueStore).

Signed-off-by: Sage Weil <sage@redhat.com>