osd: osd read object's attrs failed at EC recovery
In EC recovery read, only one shard will read the 'attrs' of object, and we will not try to read it again
when it failed which may be caused by aio failure for example. We need to read the object's 'attrs' again
if we don't have it, when trying to send all remaining reads.

Signed-off-by: xiaofei cui <cuixiaofei@sangfor.com>