osd: read object attrs but with errors crash the primary osd
In EC recovery read, if we read attrs but with errors, the primary
osd will assert. To avoid this, we need erase the attrs if there
is an reading error.
This can be reproduced by: down primary osd, put data to the pg,
inject mdata err to the new primary osd, restart the old primary osd.
The assert can be seen at the old primary osd's log.

Signed-off-by: xiaofei cui <cuixiaofei@sangfor.com>