build: mgr: check for python's ssl version linkage
If multiple openssl versions are present, it can so happen that system python
library is linked to an old openssl and openssl development headers point to a
new openssl. This would later cause segmentation fault for ceph-mgr when ssl
module is used for commands like self signed certificate as ceph-mgr itself
would've loaded a different openssl version before due to the recent msgr change
moving from NSS to openssl. Detect this condition at build time so this
incompatibility can be fixed

Fixes: http://tracker.ceph.com/issues/24282
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>